feat: 유저 정보값 저장해서 header 리렌더링 방지
